database/requirements.txt
psycopg2-binary>=2.9,<3
sqlalchemy>=2.0,<3
python-dotenv>=1.0,<2
